+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added Isekai Succubus Heritage (Tiefling Heritage).
 ### Changed
-- Buffed Edge Lord's "Supersonic Combat" to allow for qualifying for feats that require strength.
+- Buffed Edge Lord's "Supersonic Combat" to allow for use in CMB and to qualify for feats that require strength.
 
 ## [1.4.1] - 2022-11-05
 ### Fixed